# backend-laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dokumentasi API",
        "description": "Dokumentasi API v1.0.0",
        "contact": {
            "email": "prasdimasetyo@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/jobs": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Get a list of all jobs",
                "operationId": "9538d30d530d4c3680ea31fe228944e2",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": [
                                    {
                                        "id": 1,
                                        "name": "Frontend Web Developer",
                                        "created_by": null,
                                        "created_at": "2024-02-20T14:52:29.000000Z",
                                        "updated_by": null,
                                        "updated_at": "2024-02-20T14:52:29.000000Z",
                                        "deleted_by": null,
                                        "deleted_at": null,
                                        "candidates": []
                                    },
                                    {
                                        "id": 2,
                                        "name": "Fullstack Web Developer",
                                        "created_by": null,
                                        "created_at": "2024-02-20T14:52:29.000000Z",
                                        "updated_by": null,
                                        "updated_at": "2024-02-20T14:52:29.000000Z",
                                        "deleted_by": null,
                                        "deleted_at": null,
                                        "candidates": []
                                    },
                                    {
                                        "id": 3,
                                        "name": "Quality Control",
                                        "created_by": null,
                                        "created_at": "2024-02-20T14:52:29.000000Z",
                                        "updated_by": null,
                                        "updated_at": "2024-02-20T14:52:29.000000Z",
                                        "deleted_by": null,
                                        "deleted_at": null,
                                        "candidates": []
                                    }
                                ]
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Create a new job",
                "operationId": "36deb2a6480382cf8385911c10aeebf0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "name": "New Job Title"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Job created successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "job": {
                                        "id": 2,
                                        "name": "New Job Title"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "object",
                                            "example": {
                                                "name": [
                                                    "The name field is required."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/jobs/{id}": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Get job details by ID",
                "operationId": "6de31489b41fa51255446ac2ad15b7fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "id": 1,
                                    "name": "Frontend Web Developer",
                                    "created_by": null,
                                    "created_at": "2024-02-20T14:52:29.000000Z",
                                    "updated_by": null,
                                    "updated_at": "2024-02-20T14:52:29.000000Z",
                                    "deleted_by": null,
                                    "deleted_at": null,
                                    "candidates": []
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Job not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Update an existing job",
                "operationId": "66762b778fbd20dd4fbd09628dabba17",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "name": "Updated Job Title"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Job updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "job": {
                                        "id": 2,
                                        "name": "Updated Job Title"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "object",
                                            "example": {
                                                "name": [
                                                    "The name field must be a string."
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Job not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Delete an existing job",
                "operationId": "5537fd998abc3a62eef8c09e4ae52000",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Job ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Job deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Job deleted successfully"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Job not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}